Create `GrpcUtil` class
