+++ grakn-test-tools/src/main/java/ai/grakn/engine/rpc/GrpcTestUtil.java
@@ -30,7 +30,7 @@ import static org.hamcrest.Matchers.isA;
 /**
  * @author Felix Chapman
  */
-public class GrpcUtil {
+public class GrpcTestUtil {
 
     public static Matcher<StatusRuntimeException> hasStatus(Status status) {
